# v2/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Vestorly Developers API",
    "version" : "1.0.0",
    "title" : "Vestorly API",
    "termsOfService" : "http://www.vestorly.com/terms/",
    "contact" : {
      "name" : "Vestorly Team"
    },
    "license" : {
      "name" : "MIT"
    }
  },
  "host" : "staging.vestorly.com",
  "basePath" : "/api/v2",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/articles" : {
      "get" : {
        "tags" : [ "articles" ],
        "description" : "Returns all articles",
        "operationId" : "findArticles",
        "parameters" : [ {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "Limit on the number of articles to return",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "text_query",
          "in" : "query",
          "description" : "Search query parameter",
          "required" : false,
          "type" : "string",
          "format" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Response"
          }
        }
      }
    },
    "/articles/{id}" : {
      "get" : {
        "tags" : [ "articles" ],
        "description" : "Returns a single article",
        "operationId" : "findArticleByID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Article Id to fetch",
          "required" : true,
          "type" : "string",
          "format" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Article response",
            "schema" : {
              "$ref" : "#/definitions/Article"
            }
          }
        }
      }
    },
...
    "/members" : {
      "get" : {
        "tags" : [ "members" ],
        "description" : "Returns all members",
        "operationId" : "findMembers",
        "parameters" : [ {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Member response",
            "schema" : {
              "$ref" : "#/definitions/Members"
            }
          }
        }
      }
    },
    "/members/{id}" : {
      "get" : {
        "tags" : [ "members" ],
        "description" : "Returns a single member",
        "operationId" : "findMemberByID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Mongo ID of member to fetch",
          "required" : true,
          "type" : "string",
          "format" : "string"
        }, {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Member response",
            "schema" : {
              "$ref" : "#/definitions/Member"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "members" ],
        "description" : "Returns a single member",
        "operationId" : "findMemberByID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Mongo ID of member to fetch",
          "required" : true,
          "type" : "string",
          "format" : "string"
        }, {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Member response",
            "schema" : {
              "$ref" : "#/definitions/Member"
            }
          }
        }
      }
    },
    "/newsletter_settings" : {
      "get" : {
        "tags" : [ "newsletter_settings" ],
        "description" : "Returns all newsletter settings",
        "operationId" : "findNewsletterSettings",
        "parameters" : [ {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "newsletter settings response",
            "schema" : {
              "$ref" : "#/definitions/NewsletterSettings"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "newsletter_settings" ],
        "description" : "Update a single newsletter setting by ID",
        "operationId" : "updateNewsletterSettingsByID",
        "parameters" : [ {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        }, {
          "name" : "newsletter_settings",
          "in" : "formData",
          "description" : "newsletter settings",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "newsletter settings response",
            "schema" : {
              "$ref" : "#/definitions/NewsletterSettings"
            }
          }
        }
      }
    },
    "/newsletter_settings/{id}" : {
      "get" : {
        "tags" : [ "newsletter_settings" ],
        "description" : "Returns a single newsletter settings if the user has access",
        "operationId" : "findNewsletterSettingsByID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Mongo ID of newsletter settings to fetch",
          "required" : true,
          "type" : "string",
          "format" : "string"
        }, {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "newsletting setting response",
            "schema" : {
              "$ref" : "#/definitions/NewsletterSettings"
            }
          }
        }
      }
    },
    "/newsletters" : {
      "get" : {
        "tags" : [ "newsletters" ],
        "description" : "Returns all newsletters",
        "operationId" : "findNewsletters",
        "parameters" : [ {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Response",
            "schema" : {
              "$ref" : "#/definitions/Newsletters"
            }
          }
        }
      }
    },
    "/newsletters/{id}" : {
      "get" : {
        "tags" : [ "newsletters" ],
        "description" : "Returns newsletter by ID",
        "operationId" : "getNewsletterByID",
        "parameters" : [ {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "ID of newsletter to fetch",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Response",
            "schema" : {
              "$ref" : "#/definitions/Newsletter"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "newsletters" ],
        "description" : "Update newsletter by ID",
        "operationId" : "updateNewsletter",
        "parameters" : [ {
          "name" : "vestorly-auth",
          "in" : "query",
          "description" : "Vestorly Auth Token",
          "required" : true,
          "type" : "string",
          "format" : "string"
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "ID of newsletter to fetch",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Newsletter",
          "in" : "formData",
          "description" : "Newsletter",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Response",
            "schema" : {
              "$ref" : "#/definitions/Newsletter"
            }
          }
        }
      }
    },
...
    "/sessions" : {
      "post" : {
        "tags" : [ "sessions" ],
        "description" : "Login To Vestorly Platform",
        "operationId" : "login",
        "consumes" : [ "application/x-www-form-urlencoded" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "formData",
          "description" : "Username in the vestorly platform",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "formData",
          "description" : "Password in Vestorly Platform",
          "required" : true,
          "type" : "password"
        } ],
        "responses" : {
          "200" : {
            "description" : "Response",
            "schema" : {
              "$ref" : "#/definitions/Session"
            }
          }
        }
      }
    },
    "/sessions/{id}" : {
      "delete" : {
        "tags" : [ "sessions" ],
        "description" : "Logout",
        "operationId" : "logout",
        "consumes" : [ "application/x-www-form-urlencoded" ],
        "parameters" : [ {
          "name" : "vestorly-auth",
          "in" : "formData",
          "description" : "Authenication token",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Response",
            "schema" : {
              "$ref" : "#/definitions/Session"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Members" : {
      "properties" : {
        "members" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Member"
          }
        }
      }
    },
...
    "Article" : { },
...
    "Newsletter" : {
      "properties" : {
        "is_sent" : {
          "type" : "boolean",
          "default" : false
        },
        "is_default" : {
          "type" : "boolean",
          "default" : false
        },
        "click_count" : {
          "type" : "integer",
          "format" : "int64"
        },
        "unique_click_count" : {
          "type" : "integer",
          "format" : "int64"
        },
        "total_click_count" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
...
    "Newsletters" : {
      "properties" : {
        "newletters" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Newsletter"
          }
        }
      }
    },
...
    "NewsletterSettings" : {
      "required" : [ "_id" ],
      "properties" : {
        "_id" : {
          "type" : "string"
        },
        "email_status" : {
          "type" : "string"
        },
        "twitter_status" : {
          "type" : "string"
        },
        "linkedin_status" : {
          "type" : "string"
        },
        "email_hour" : {
          "type" : "integer",
          "format" : "int32"
        },
        "twitter_hour" : {
          "type" : "integer",
          "format" : "int32"
        },
        "linkedin_hour" : {
          "type" : "integer",
          "format" : "int32"
        },
        "facebook_hour" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    }
  }
}